remove things from init
